+++ spec/article_json/article_spec.rb
@@ -53,6 +53,17 @@ describe ArticleJSON::Article do
     it { should eq '<p>Foo Bar</p>' }
   end
 
+  describe '#facebook_instant_article_exporter' do
+    subject { article.facebook_instant_article_exporter }
+    it { should be_a ArticleJSON::Export::FacebookInstantArticle::Exporter }
+  end
+
+  describe '#to_facebook_instant_article' do
+    subject { article.to_facebook_instant_article }
+    it { should be_a String }
+    it { should eq '<p>Foo Bar</p>' }
+  end
+
   describe '#plain_text_exporter' do
     subject { article.plain_text_exporter }
     it { should be_a ArticleJSON::Export::PlainText::Exporter }